+++ src/lib.rs
@@ -8,6 +8,7 @@ mod http_languages;
 mod http_process;
 mod ip_options;
 mod mtu;
+mod observable_signals;
 pub mod p0f_output;
 mod process;
 mod signature_matcher;
