+++ src/observable_signals.rs
@@ -1,0 +1,62 @@
+use crate::http::{Header, Version};
+use crate::tcp::{IpVersion, PayloadSize, Quirk, TcpOption, WindowSize};
+use crate::Ttl;
+
+#[derive(Debug, Clone)]
+pub struct ObservableTcp {
+    pub version: IpVersion,
+    /// initial TTL used by the OS.
+    pub ittl: Ttl,
+    /// length of IPv4 options or IPv6 extension headers.
+    pub olen: u8,
+    /// maximum segment size, if specified in TCP options.
+    pub mss: Option<u16>,
+    /// window size.
+    pub wsize: WindowSize,
+    /// window scaling factor, if specified in TCP options.
+    pub wscale: Option<u8>,
+    /// layout and ordering of TCP options, if any.
+    pub olayout: Vec<TcpOption>,
+    /// properties and quirks observed in IP or TCP headers.
+    pub quirks: Vec<Quirk>,
+    /// payload size classification
+    pub pclass: PayloadSize,
+}
+
+#[derive(Debug, Clone)]
+pub struct ObservableHttpRequest {
+    pub lang: Option<String>,
+    pub user_agent: Option<String>,
+    /// HTTP version
+    pub version: Version,
+    /// ordered list of headers that should appear in matching traffic.
+    pub horder: Vec<Header>,
+    /// list of headers that must *not* appear in matching traffic.
+    pub habsent: Vec<Header>,
+    /// expected substring in 'User-Agent' or 'Server'.
+    pub expsw: String,
+}
+
+#[derive(Debug, Clone)]
+pub struct ObservableHttpResponse {
+    /// HTTP version
+    pub version: Version,
+    /// ordered list of headers that should appear in matching traffic.
+    pub horder: Vec<Header>,
+    /// list of headers that must *not* appear in matching traffic.
+    pub habsent: Vec<Header>,
+    /// expected substring in 'User-Agent' or 'Server'.
+    pub expsw: String,
+}
+
+pub struct ObservableMtu {
+    pub value: u16,
+}
+
+pub struct ObservableUptime {
+    pub days: u32,
+    pub hours: u32,
+    pub min: u32,
+    pub up_mod_days: u32,
+    pub freq: f64,
+}
