+++ src/p0f_output.rs
@@ -1,9 +1,9 @@
 use crate::db::{Label, Type};
 use crate::http::HttpDiagnosis;
-use crate::http_process::{ObservableHttpRequest, ObservableHttpResponse};
+use crate::observable_signals::ObservableTcp;
+use crate::observable_signals::{ObservableHttpRequest, ObservableHttpResponse};
 use crate::process::IpPort;
 use crate::tcp::Ttl;
-use crate::tcp_process::ObservableTcp;
 use std::fmt;
 use std::fmt::Formatter;
